integer overflow fixed

+++ code/trunk/src/util/Math.h
@@ -219,7 +219,7 @@ T interpolateSmooth(float time, const T& start, const T& end)
 */
 inline _UtilExport float rnd()
 {
-    return ((float)rand() / (RAND_MAX + 1));
+    return rand() / (RAND_MAX + 1.0);
 }
 
 /**
